Minor restructure, moving corpus and ngram code into their own modules
